Organize topology layout
